+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Please use the following tags when editing this file:
 *Removed* for now removed features.
 *Fixed* for any bug fixes. 
 
+## Version 3.3
+### Changed
+- Simplified UI. Now output directory and prefix are specified independently.
+  NB: This breaks CLI backward compatibility
+- Treat a relative output directory as relative to the original SGV file.
+
 ## Version 3.2
 ### Added
 - Enable sub-layer support, treat sub-layers in the same way as groups